Refactored mode facilities
- Refactor mode facilitites
- Mode: Added is_single_precision()
- Docs: Updated top-level documentation outline
